Add WanAccessTypeAwareFritzServiceViewModel
