+++ RS.Fritz.Manager.UI/UserControls/WanIpConnection/WanIpConnectionViewModel.cs
@@ -2,12 +2,10 @@
 
 using System.Threading;
 using System.Threading.Tasks;
-using CommunityToolkit.Mvvm.Messaging;
-using CommunityToolkit.Mvvm.Messaging.Messages;
 using Microsoft.Extensions.Logging;
 using RS.Fritz.Manager.API;
 
-internal sealed class WanIpConnectionViewModel : FritzServiceViewModel, IRecipient<PropertyChangedMessage<WanAccessType?>>
+internal sealed class WanIpConnectionViewModel : WanAccessTypeAwareFritzServiceViewModel
 {
     private WanIpConnectionGetInfoResponse? wanIpConnectionGetInfoResponse;
     private WanIpConnectionGetConnectionTypeInfoResponse? wanIpConnectionGetConnectionTypeInfoResponse;
@@ -18,7 +16,7 @@ internal sealed class WanIpConnectionViewModel : FritzServiceViewModel, IRecipie
     private WanIpConnectionGetExternalIpAddressResponse? wanIpConnectionGetExternalIpAddressResponse;
 
     public WanIpConnectionViewModel(DeviceLoginInfo deviceLoginInfo, ILogger logger)
-        : base(deviceLoginInfo, logger)
+        : base(deviceLoginInfo, logger, WanAccessType.Ethernet)
     {
     }
 
@@ -64,21 +62,6 @@ internal sealed class WanIpConnectionViewModel : FritzServiceViewModel, IRecipie
         private set { _ = SetProperty(ref wanIpConnectionGetExternalIpAddressResponse, value); }
     }
 
-    public void Receive(PropertyChangedMessage<WanAccessType?> message)
-    {
-        if (message.Sender != DeviceLoginInfo.InternetGatewayDevice)
-            return;
-
-        switch (message.PropertyName)
-        {
-            case nameof(ObservableInternetGatewayDevice.WanAccessType):
-                {
-                    UpdateCanExecuteDefaultCommand();
-                    break;
-                }
-        }
-    }
-
     protected override async Task DoExecuteDefaultCommandAsync(CancellationToken cancellationToken = default)
     {
         await API.TaskExtensions.WhenAllSafe(new[]
@@ -91,11 +74,6 @@ internal sealed class WanIpConnectionViewModel : FritzServiceViewModel, IRecipie
                 GetWanIpConnectionGetPortMappingNumberOfEntriesAsync(),
                 GetWanIpConnectionGetExternalIpAddressAsync()
           });
-    }
-
-    protected override bool GetCanExecuteDefaultCommand()
-    {
-        return base.GetCanExecuteDefaultCommand() && DeviceLoginInfo.InternetGatewayDevice!.WanAccessType is WanAccessType.Ethernet;
     }
 
     private async Task GetWanIpConnectionGetInfoAsync()
